+++ src/Scissors.Test/RankedRoutes/RankedRouteFixture.cs
@@ -1,6 +1,6 @@
-﻿using static Scissors.HttpRequestInterceptor.Test.InMemoryConfigurationWriter;
+﻿using static Scissors.Test.InMemoryConfigurationWriter;
 
-namespace Scissors.HttpRequestInterceptor.Test.RankedRoutes
+namespace Scissors.Test.RankedRoutes
 {
     public class RankedRouteFixture : TestFixture
     {