+++ src/Scissors.Test/ResponseContents/ResponseContentTests.cs
@@ -7,7 +7,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace Scissors.HttpRequestInterceptor.Test.ResponseContents
+namespace Scissors.Test.ResponseContents
 {
     public class ResponseContentTests : IClassFixture<ResponseContentFixture>
     {